refactored depth layer into separate file
